+++ mne/viz/backends/_pyvista.py
@@ -22,8 +22,8 @@ import numpy as np
 import vtk
 
 from .abstract_renderer import _AbstractRenderer
-from .abstract_dock import _AbstractDock
-from .abstract_tool_bar import _AbstractToolBar
+from ._qt_dock import _QtDock
+from ._qt_tool_bar import _QtToolBar
 from ._utils import (_get_colormap_from_array, _alpha_blend_background,
                      ALLOWED_QUIVER_MODES, _init_qt_resources,
                      _qt_disable_paint)
@@ -856,181 +856,9 @@ class _PyVistaRenderer(_AbstractRenderer):
             prop.SetLineWidth(line_width)
 
 
-class _Renderer(_PyVistaRenderer, _AbstractDock, _AbstractToolBar):
-    def _dock_initialize(self):
-        from PyQt5 import QtCore
-        from PyQt5.QtWidgets import QDockWidget, QVBoxLayout, QWidget
-        self.dock = QDockWidget()
-        self.dock.setAllowedAreas(QtCore.Qt.LeftDockWidgetArea)
-        self.dock.setFeatures(QDockWidget.NoDockWidgetFeatures)
-        self.plotter.app_window.addDockWidget(
-            QtCore.Qt.LeftDockWidgetArea, self.dock)
-        self.dock.setWidget(QWidget())
-        self.dock_layout = QVBoxLayout()
-        self.dock.widget().setLayout(self.dock_layout)
-
-    def _dock_finalize(self):
-        self._dock_add_stretch(self.dock_layout)
-
-    def _dock_show(self):
-        self.dock.show()
-
-    def _dock_hide(self):
-        self.dock.hide()
-
-    def _dock_add_stretch(self, layout):
-        layout.addStretch()
-
-    def _dock_add_layout(self, vertical=True):
-        from PyQt5.QtWidgets import QHBoxLayout, QVBoxLayout
-        layout = QVBoxLayout() if vertical else QHBoxLayout()
-        return layout
-
-    def _dock_add_label(self, value, align=False, layout=None):
-        from PyQt5 import QtCore
-        from PyQt5.QtWidgets import QLabel
-        layout = self.dock_layout if layout is None else layout
-        widget = QLabel()
-        if align:
-            widget.setAlignment(QtCore.Qt.AlignCenter)
-        widget.setText(value)
-        layout.addWidget(widget)
-        return widget
-
-    def _dock_add_button(self, name, callback, layout=None):
-        from PyQt5.QtWidgets import QPushButton
-        layout = self.dock_layout if layout is None else layout
-        widget = QPushButton(name)
-        widget.released.connect(callback)
-        layout.addWidget(widget)
-        return widget
-
-    def _dock_add_text(self, value, callback, validator=None,
-                       layout=None):
-        from PyQt5 import QtCore
-        from PyQt5.QtGui import QDoubleValidator
-        from PyQt5.QtWidgets import QLineEdit
-        layout = self.dock_layout if layout is None else layout
-        widget = QLineEdit(value)
-        widget.setAlignment(QtCore.Qt.AlignCenter)
-        if validator is not None:
-            widget.setValidator(
-                QDoubleValidator(validator[0], validator[1], 2))
-
-            def _callback():
-                callback(float(widget.text()))
-        else:
-            def _callback():
-                callback(widget.text())
-        widget.returnPressed.connect(_callback)
-        layout.addWidget(widget)
-        return widget
-
-    def _dock_add_slider(self, name, value, rng, callback,
-                         compact=True, double=False, layout=None):
-        from PyQt5 import QtCore
-        from PyQt5.QtWidgets import QSlider
-        from .float_slider import float_slider_class
-        layout = self.dock_layout if layout is None else layout
-        hlayout = self._dock_add_layout(not compact)
-        if name is not None:
-            self._dock_add_label(
-                value=name, align=not compact, layout=hlayout)
-        slider_class = float_slider_class() if double else QSlider
-        widget = slider_class(QtCore.Qt.Horizontal)
-        widget.setMinimum(rng[0] if double else int(rng[0]))
-        widget.setMaximum(rng[1] if double else int(rng[1]))
-        widget.setValue(value if double else int(value))
-        widget.valueChanged.connect(callback)
-        hlayout.addWidget(widget)
-        layout.addLayout(hlayout)
-        return widget
-
-    def _dock_add_spin_box(self, name, value, rng, callback,
-                           compact=True, double=True, layout=None):
-        from PyQt5 import QtCore
-        from PyQt5.QtWidgets import QDoubleSpinBox, QSpinBox
-        layout = self.dock_layout if layout is None else layout
-        hlayout = self._dock_add_layout(not compact)
-        if name is not None:
-            self._dock_add_label(
-                value=name, align=not compact, layout=hlayout)
-        value = value if double else int(value)
-        widget = QDoubleSpinBox() if double else QSpinBox()
-        widget.setAlignment(QtCore.Qt.AlignCenter)
-        widget.setMinimum(rng[0])
-        widget.setMaximum(rng[1])
-        widget.setValue(value)
-        widget.valueChanged.connect(callback)
-        hlayout.addWidget(widget)
-        layout.addLayout(hlayout)
-        return widget
-
-    def _dock_add_combo_box(self, name, value, rng,
-                            callback, compact=True, layout=None):
-        from PyQt5.QtWidgets import QComboBox
-        layout = self.dock_layout if layout is None else layout
-        hlayout = self._dock_add_layout(not compact)
-        if name is not None:
-            self._dock_add_label(
-                value=name, align=not compact, layout=hlayout)
-        widget = QComboBox()
-        widget.addItems(rng)
-        widget.setCurrentText(value)
-        widget.currentTextChanged.connect(callback)
-        hlayout.addWidget(widget)
-        layout.addLayout(hlayout)
-        return widget
-
-    def _dock_add_group_box(self, name, layout=None):
-        from PyQt5.QtWidgets import QGroupBox, QVBoxLayout
-        layout = self.dock_layout if layout is None else layout
-        hlayout = QVBoxLayout()
-        widget = QGroupBox(name)
-        widget.setLayout(hlayout)
-        layout.addWidget(widget)
-        return hlayout
-
-    def _tool_bar_load_icons(self):
-        from PyQt5.QtGui import QIcon
-        from ._utils import _init_qt_resources
-        _init_qt_resources()
-        self.icons = dict()
-        self.icons["help"] = QIcon(":/help.svg")
-        self.icons["play"] = QIcon(":/play.svg")
-        self.icons["pause"] = QIcon(":/pause.svg")
-        self.icons["reset"] = QIcon(":/reset.svg")
-        self.icons["scale"] = QIcon(":/scale.svg")
-        self.icons["clear"] = QIcon(":/clear.svg")
-        self.icons["movie"] = QIcon(":/movie.svg")
-        self.icons["restore"] = QIcon(":/restore.svg")
-        self.icons["screenshot"] = QIcon(":/screenshot.svg")
-        self.icons["visibility_on"] = QIcon(":/visibility_on.svg")
-        self.icons["visibility_off"] = QIcon(":/visibility_off.svg")
-
-    def _tool_bar_initialize(self, name="default"):
-        self.actions = dict()
-        self.tool_bar = self.plotter.app_window.addToolBar(name)
-
-    def _tool_bar_finalize(self):
-        pass
-
-    def _tool_bar_add_button(self, name, desc, func, icon_name=None):
-        icon_name = name if icon_name is None else icon_name
-        icon = self.icons[icon_name]
-        self.actions[name] = self.tool_bar.addAction(icon, desc, func)
-
-    def _tool_bar_update_button_icon(self, name, icon_name):
-        self.actions[name].setIcon(self.icons[icon_name])
-
-    def _tool_bar_add_text(self, name, value, placeholder):
-        pass
-
-    def _tool_bar_add_spacer(self):
-        from PyQt5.QtWidgets import QSizePolicy, QWidget
-        spacer = QWidget()
-        spacer.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Preferred)
-        self.tool_bar.addWidget(spacer)
+class _Renderer(_PyVistaRenderer, _QtDock, _QtToolBar):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 
 
 def _compute_normals(mesh):
